Changes:
- fixed here doc when it didnt stop when reached delimited;
