+++ src/executor/executor.c
@@ -6,7 +6,7 @@
 /*   By: iasonov <iasonov@student.42prague.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/17 00:40:18 by iasonov           #+#    #+#             */
-/*   Updated: 2025/02/14 21:20:05 by iasonov          ###   ########.fr       */
+/*   Updated: 2025/02/15 20:46:46 by iasonov          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
